Back-end:
Added address and phone fields to clientEntity.
Added totalItems field to BillEntity. modify the "update()" method inside the clientController and the "create()" method also.
Modify the "Update"

Front-end:

update BillCreateComponent, BillDetailComponent in order to show the number of total ítems in the Bill.
update ProductDetailComponent to show in a better way the image of the product.
update Bill detail view to show small image of bought product.
added phone and address fields to the view for create clients
added  phone and address fields to the update view for clients
fix problem with the form on the update and create product views.
added image to "update" product view.
added extra field to upload product image in the "create" and "update" views.